add request and response dto to GetData controller
